+++ RoboMinerCpp/Robot.h
@@ -45,7 +45,7 @@ public:
 public:
     CRobot(int maxTurns, int maxOre,
            int miningSpeed, int cpuSpeed,
-           int forwardSpeed, int backwardSpeed, int rotateSpeed,
+           double forwardSpeed, double backwardSpeed, int rotateSpeed,
            int robotSize);
     virtual ~CRobot();
